frontend/requirements_frontend.txt
streamlit>=1.28.0
streamlit-authenticator>=0.3.3
bcrypt>=4.0.1
PyYAML>=6.0
pandas>=2.0.0
plotly>=5.17.0
